+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.2 ##
+
+Add XigniteGlobalHistorical > GetEndOfDayQuotes endpoint.
+
+## 0.2.1 ##
+
+Add XigniteGlobalMaster > GetMasterByExchange endpoint.
+
 ## 0.1.2 ##
 
 Adds support and CI builds for both Ruby 1.8.7 and 1.9.2.